Move non-view files to non-view package
